Refactored Food APIs
