+++ Food Ordering App Backend/schemas/food.js
@@ -28,7 +28,7 @@ const foodSchema = new mongoose.Schema({
 
     addedAt : {
         type : Date,
-        default: () => Date.now()        
+        default: Date.now()        
     }
 
 });
