+++ Food Ordering App Backend/schemas/order.js
@@ -34,7 +34,7 @@ const cartSchema = new mongoose.Schema({
 
     addedAt : {
         type : Date,
-        default: () => Date.now()        
+        default: Date.now()        
     }
 
 });
