+++ Food Ordering App Backend/schemas/restaurant.js
@@ -19,7 +19,7 @@ const restaurantSchema = new mongoose.Schema({
 
     addedAt : {
         type : Date,
-        default: () => Date.now()        
+        default: Date.now()        
     }
 });
