+++ Food Ordering App Backend/schemas/user.js
@@ -47,7 +47,7 @@ const userSchema = new mongoose.Schema({
 
     createdAt : {
         type : Date,
-        default: () => Date.now()
+        default: Date.now()
     },
 
     token : {
